Move the console-related code to a dedicated module.
