+++ src/client/logging/logger.ts
@@ -2,14 +2,15 @@
 // Licensed under the MIT License.
 'use strict';
 
-// tslint:disable:no-console no-any
+// tslint:disable:no-any
 
 import * as util from 'util';
 import { createLogger } from 'winston';
 import { isTestExecution } from '../common/constants';
+import { logToConsole, monkeypatchConsole } from './_console';
 import { getFormatter } from './formatters';
 import { getConsoleTransport, getFileTransport } from './transports';
-import { ConsoleStreams, LogLevel } from './types';
+import { LogLevel } from './types';
 
 // Initialize the loggers as soon as this module is imported.
 const consoleLogger = createLogger();
@@ -42,23 +43,6 @@ function logToFile(logLevel: LogLevel, ...args: any[]) {
     fileLogger.log(logLevelMap[logLevel], message);
 }
 
-const logMethods = {
-    log: Symbol.for('log'),
-    info: Symbol.for('info'),
-    error: Symbol.for('error'),
-    debug: Symbol.for('debug'),
-    warn: Symbol.for('warn')
-};
-
-function logToConsole(stream: ConsoleStreams, ...args: any[]) {
-    if (['info', 'error', 'warn', 'log', 'debug'].indexOf(stream) === -1) {
-        stream = 'log';
-    }
-    // Further below we monkeypatch the console.log, etc methods.
-    const fn = (console as any)[logMethods[stream]] || console[stream] || console.log;
-    fn(...args);
-}
-
 /**
  * Initialize the logger for console.
  * We do two things here:
@@ -74,53 +58,10 @@ function logToConsole(stream: ConsoleStreams, ...args: any[]) {
  *   To do this we need to monkey patch the console methods.
  *   This is optional (generally done when running tests on CI).
  */
-// tslint:disable-next-line: max-func-body-length
 function initializeConsoleLogger() {
     // Hijack `console.log` when running tests on CI.
     if (process.env.VSC_PYTHON_LOG_FILE && process.env.TF_BUILD) {
-        /*
-        What we're doing here is monkey patching the console.log so we can send everything sent to console window into our logs.
-        This is only required when we're directly writing to `console.log` or not using our `winston logger`.
-        This is something we'd generally turn on, only on CI so we can see everything logged to the console window (via the logs).
-        */
-        // Keep track of the original functions before we monkey patch them.
-        // Using symbols guarantee the properties will be unique & prevents clashing with names other code/library may create or have created.
-        (console as any)[logMethods.log] = console.log;
-        (console as any)[logMethods.info] = console.info;
-        (console as any)[logMethods.error] = console.error;
-        (console as any)[logMethods.debug] = console.debug;
-        (console as any)[logMethods.warn] = console.warn;
-
-        // tslint:disable-next-line: no-function-expression
-        console.log = function () {
-            const args = Array.prototype.slice.call(arguments);
-            logToConsole('log', ...args);
-            logToFile(LogLevel.Information, ...args);
-        };
-        // tslint:disable-next-line: no-function-expression
-        console.info = function () {
-            const args = Array.prototype.slice.call(arguments);
-            logToConsole('info', ...args);
-            logToFile(LogLevel.Information, ...args);
-        };
-        // tslint:disable-next-line: no-function-expression
-        console.warn = function () {
-            const args = Array.prototype.slice.call(arguments);
-            logToConsole('warn', ...args);
-            logToFile(LogLevel.Warning, ...args);
-        };
-        // tslint:disable-next-line: no-function-expression
-        console.error = function () {
-            const args = Array.prototype.slice.call(arguments);
-            logToConsole('error', ...args);
-            logToFile(LogLevel.Error, ...args);
-        };
-        // tslint:disable-next-line: no-function-expression
-        console.debug = function () {
-            const args = Array.prototype.slice.call(arguments);
-            logToConsole('debug', ...args);
-            logToFile(LogLevel.Information, ...args);
-        };
+        monkeypatchConsole(logToFile);
     }
 
     if (isTestExecution() && !process.env.VSC_PYTHON_FORCE_LOGGING) {
